+++ packages/banner/src/css/index.ts
@@ -12,7 +12,7 @@ import {
 } from '@pluralsight/ps-design-system-core'
 import { transparentize } from '@pluralsight/ps-design-system-util'
 
-import * as vars from '../vars'
+import * as vars from '../vars/index.js'
 
 export default {
   '.psds-banner': {
